Refactored fluorophore commands directory layout.
