+++ scripts/example_random_3d_fluorophores.bsh
@@ -15,7 +15,7 @@ import ch.epfl.leb.sass.models.obstructors.internal.commands.GenerateFiducialsRa
 import ch.epfl.leb.sass.models.backgrounds.internal.commands.GenerateUniformBackground;
 import ch.epfl.leb.sass.models.components.internal.*;
 import ch.epfl.leb.sass.models.psfs.internal.Gaussian3D;
-import ch.epfl.leb.sass.models.fluorophores.internal.commands.GenerateFluorophoresRandom3D;
+import ch.epfl.leb.sass.models.fluorophores.commands.internal.GenerateFluorophoresRandom3D;
 import ch.epfl.leb.sass.utils.RNG;
 
 // The seed determines the outputs of the random number generator.
