+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/commands/internal/FluorophoreReceiver.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.models.fluorophores.internal.commands;
+package ch.epfl.leb.sass.models.fluorophores.commands.internal;
 
 import ch.epfl.leb.sass.utils.RNG;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
